Corrects checksum addresses in main function file.

+++ main_function_ino/main_function_ino.ino
@@ -2,11 +2,11 @@
 #include <Servo.h>                                            // Calls for Servo library
 
 #define MD25ADDRESS         0x58                              // Address of the MD25
-#define SPEED1              (byte) 0x00                              // Byte to send speed to both motors for forward and backwards motion if operated in MODE 2 or 3 and Motor 1 Speed if in MODE 0 or 1
+#define SPEED1              0x00                              // Byte to send speed to both motors for forward and backwards motion if operated in MODE 2 or 3 and Motor 1 Speed if in MODE 0 or 1
 #define SPEED2              0x01                              // Byte to send speed for turn speed if operated in MODE 2 or 3 and Motor 2 Speed if in MODE 0 or 1
 #define ENCODERONE          0x02                              // Byte to read motor encoder 1
 #define ENCODERTWO          0x06                              // Byte to read motor encoder 2
-#define ACCELERATION        0xE                               // Byte to def===ine motor acceleration
+#define ACCELERATION        0xE                               // Byte to define motor acceleration
 #define CMD                 0x10                              // Byte to reset encoder values
 #define MODE_SELECTOR       0xF                               // Byte to change between control MODES
